requirements.txt
discord.py>=2.3.2
python-dateutil>=2.8.2
pytz>=2023.3
apscheduler>=3.10.1
python-dotenv>=1.0.0